+++ openrl/utils/callbacks/callbacks.py
@@ -5,23 +5,13 @@ from abc import ABC, abstractmethod
 from typing import Any, Callable, Dict, List, Optional, Union
 
 import gym
-import numpy as np
 
 from openrl.utils.logger import Logger
 
-try:
-    from tqdm import TqdmExperimentalWarning
-
-    # Remove experimental warning
-    warnings.filterwarnings("ignore", category=TqdmExperimentalWarning)
-    from tqdm.rich import tqdm
-except ImportError:
-    # Rich not installed, we only throw an error
-    # if the progress bar is used
-    tqdm = None
 
 from openrl.envs.vec_env import BaseVecEnv
 from openrl.runners.common.base_agent import BaseAgent
+import openrl.utils.callbacks.callbacks_factory as callbacks_factory
 
 
 class BaseCallback(ABC):
@@ -134,6 +124,14 @@ class BaseCallback(ABC):
         """
         pass
 
+    def set_parent(self, parent: "BaseCallback") -> None:
+        """
+        Set the parent of the callback.
+
+        :param parent: The parent callback.
+        """
+        self.parent = parent
+
 
 class EventCallback(BaseCallback):
     """
@@ -149,7 +147,7 @@ class EventCallback(BaseCallback):
         self.callback = callback
         # Give access to the parent
         if callback is not None:
-            self.callback.parent = self
+            self.callback.set_parent(self)
 
     def init_callback(self, agent: "BaseAgent") -> None:
         super().init_callback(agent)
@@ -186,10 +184,11 @@ class CallbackList(BaseCallback):
         sequentially.
     """
 
-    def __init__(self, callbacks: List[BaseCallback]):
+    def __init__(self, callbacks: List[BaseCallback], stop_logic: str = "OR"):
         super().__init__()
         assert isinstance(callbacks, list)
         self.callbacks = callbacks
+        self.stop_logic = stop_logic
 
     def _init_callback(self) -> None:
         for callback in self.callbacks:
@@ -204,11 +203,28 @@ class CallbackList(BaseCallback):
             callback.on_rollout_start()
 
     def _on_step(self) -> bool:
-        continue_training = True
+        if self.stop_logic == "OR":
+            # any callback return should_stop, then to stop
+            should_stop = False
+        elif self.stop_logic == "AND":
+            # all callbacks return should_stop, then to stop
+            should_stop = True
+        else:
+            raise ValueError(
+                "Unknown stop logic {}, possible values are 'OR' or 'AND'".format(
+                    self.stop_logic
+                )
+            )
+
         for callback in self.callbacks:
             # Return False (stop training) if at least one callback returns False
-            continue_training = callback.on_step() and continue_training
-        return continue_training
+            if self.stop_logic == "OR":
+                should_stop = (not callback.on_step()) or should_stop
+
+            elif self.stop_logic == "AND":
+                should_stop = (not callback.on_step()) and should_stop
+
+        return not should_stop
 
     def _on_rollout_end(self) -> None:
         for callback in self.callbacks:
@@ -233,6 +249,16 @@ class CallbackList(BaseCallback):
             callback_names.append(callback.__class__.__name__)
         return str(callback_names)
 
+    def set_parent(self, parent: "BaseCallback") -> None:
+        """
+        Set the parent of the callback.
+
+        :param parent: The parent callback.
+        """
+        self.parent = parent
+        for callback in self.callbacks:
+            callback.set_parent(parent)
+
 
 class ConvertCallback(BaseCallback):
     """
@@ -256,39 +282,6 @@ class ConvertCallback(BaseCallback):
         return True
 
 
-class StopTrainingOnRewardThreshold(BaseCallback):
-    """
-    Stop the training once a threshold in episodic reward
-    has been reached (i.e. when the model is good enough).
-
-    It must be used with the ``EvalCallback``.
-
-    :param reward_threshold:  Minimum expected reward per episode
-        to stop training.
-    :param verbose: Verbosity level: 0 for no output, 1 for indicating when training ended because episodic reward
-        threshold reached
-    """
-
-    def __init__(self, reward_threshold: float, verbose: int = 0):
-        super().__init__(verbose=verbose)
-        self.reward_threshold = reward_threshold
-
-    def _on_step(self) -> bool:
-        assert self.parent is not None, (
-            "``StopTrainingOnMinimumReward`` callback must be used "
-            "with an ``EvalCallback``"
-        )
-        # Convert np.bool_ to bool, otherwise callback() is False won't work
-        continue_training = bool(self.parent.best_mean_reward < self.reward_threshold)
-        if self.verbose >= 1 and not continue_training:
-            print(
-                "Stopping training because the mean reward"
-                f" {self.parent.best_mean_reward:.2f}  is above the threshold"
-                f" {self.reward_threshold}"
-            )
-        return continue_training
-
-
 class EveryNTimesteps(EventCallback):
     """
     Trigger a callback every ``n_steps`` timesteps
@@ -298,8 +291,17 @@ class EveryNTimesteps(EventCallback):
         when the event is triggered.
     """
 
-    def __init__(self, n_steps: int, callback: BaseCallback):
-        super().__init__(callback)
+    def __init__(
+        self,
+        n_steps: int,
+        callbacks: Union[List[Dict[str, Any]], Dict[str, Any], BaseCallback],
+        stop_logic: str = "OR",
+    ):
+        if isinstance(callbacks, list):
+            callbacks = callbacks_factory.CallbackFactory.get_callbacks(
+                callbacks, stop_logic=stop_logic
+            )
+        super().__init__(callbacks)
         self.n_steps = n_steps
         self.last_time_trigger = 0
 
@@ -308,134 +310,3 @@ class EveryNTimesteps(EventCallback):
             self.last_time_trigger = self.num_time_steps
             return self._on_event()
         return True
-
-
-class StopTrainingOnMaxEpisodes(BaseCallback):
-    """
-    Stop the training once a maximum number of episodes are played.
-
-    For multiple environments presumes that, the desired behavior is that the agent trains on each env for ``max_episodes``
-    and in total for ``max_episodes * n_envs`` episodes.
-
-    :param max_episodes: Maximum number of episodes to stop training.
-    :param verbose: Verbosity level: 0 for no output, 1 for indicating information about when training ended by
-        reaching ``max_episodes``
-    """
-
-    def __init__(self, max_episodes: int, verbose: int = 0):
-        super().__init__(verbose=verbose)
-        self.max_episodes = max_episodes
-        self._total_max_episodes = max_episodes
-        self.n_episodes = 0
-
-    def _init_callback(self) -> None:
-        # At start set total max according to number of envirnments
-        self._total_max_episodes = self.max_episodes * self.training_env.num_envs
-
-    def _on_step(self) -> bool:
-        # Check that the `dones` local variable is defined
-        assert "dones" in self.locals, (
-            "`dones` variable is not defined, please check your code next to"
-            " `callback.on_step()`"
-        )
-        self.n_episodes += np.sum(self.locals["dones"]).item()
-
-        continue_training = self.n_episodes < self._total_max_episodes
-
-        if self.verbose >= 1 and not continue_training:
-            mean_episodes_per_env = self.n_episodes / self.training_env.num_envs
-            mean_ep_str = (
-                f"with an average of {mean_episodes_per_env:.2f} episodes per env"
-                if self.training_env.num_envs > 1
-                else ""
-            )
-
-            print(
-                f"Stopping training with a total of {self.num_time_steps} steps because"
-                f" the {self.locals.get('tb_log_name')} model reached"
-                f" max_episodes={self.max_episodes}, by playing for"
-                f" {self.n_episodes} episodes {mean_ep_str}"
-            )
-        return continue_training
-
-
-class StopTrainingOnNoModelImprovement(BaseCallback):
-    """
-    Stop the training early if there is no new best model (new best mean reward) after more than N consecutive evaluations.
-
-    It is possible to define a minimum number of evaluations before start to count evaluations without improvement.
-
-    It must be used with the ``EvalCallback``.
-
-    :param max_no_improvement_evals: Maximum number of consecutive evaluations without a new best model.
-    :param min_evals: Number of evaluations before start to count evaluations without improvements.
-    :param verbose: Verbosity level: 0 for no output, 1 for indicating when training ended because no new best model
-    """
-
-    def __init__(
-        self, max_no_improvement_evals: int, min_evals: int = 0, verbose: int = 0
-    ):
-        super().__init__(verbose=verbose)
-        self.max_no_improvement_evals = max_no_improvement_evals
-        self.min_evals = min_evals
-        self.last_best_mean_reward = -np.inf
-        self.no_improvement_evals = 0
-
-    def _on_step(self) -> bool:
-        assert self.parent is not None, (
-            "``StopTrainingOnNoModelImprovement`` callback must be used with an"
-            " ``EvalCallback``"
-        )
-
-        continue_training = True
-
-        if self.n_calls > self.min_evals:
-            if self.parent.best_mean_reward > self.last_best_mean_reward:
-                self.no_improvement_evals = 0
-            else:
-                self.no_improvement_evals += 1
-                if self.no_improvement_evals > self.max_no_improvement_evals:
-                    continue_training = False
-
-        self.last_best_mean_reward = self.parent.best_mean_reward
-
-        if self.verbose >= 1 and not continue_training:
-            print(
-                "Stopping training because there was no new best model in the last"
-                f" {self.no_improvement_evals:d} evaluations"
-            )
-
-        return continue_training
-
-
-class ProgressBarCallback(BaseCallback):
-    """
-    Display a progress bar when training SB3 agent
-    using tqdm and rich packages.
-    """
-
-    def __init__(self) -> None:
-        super().__init__()
-        if tqdm is None:
-            raise ImportError(
-                "You must install tqdm and rich in order to use the progress bar"
-                " callback. "
-            )
-        self.pbar = None
-
-    def _on_training_start(self) -> None:
-        # Initialize progress bar
-        # Remove timesteps that were done in previous training sessions
-        self.pbar = tqdm(
-            total=self.locals["total_timesteps"] - self.agent.num_time_steps
-        )
-
-    def _on_step(self) -> bool:
-        # Update progress bar, we do num_envs steps per call to `env.step()`
-        self.pbar.update(self.training_env.num_envs)
-        return True
-
-    def _on_training_end(self) -> None:
-        # Flush and close progress bar
-        self.pbar.refresh()
-        self.pbar.close()
