+++ openrl/utils/callbacks/checkpoint_callback.py
@@ -70,7 +70,7 @@ class CheckpointCallback(BaseCallback):
         """
         return os.path.join(
             self.save_path,
-            f"{self.name_prefix}_{checkpoint_type}{self.num_time_steps}_steps.{extension}",
+            f"{self.name_prefix}_{checkpoint_type}{self.num_time_steps}_steps{'.' if extension else ''}{extension}",
         )
 
     def _on_step(self) -> bool:
