+++ openrl/utils/callbacks/eval_callback.py
@@ -17,7 +17,7 @@
 """"""
 import os
 import warnings
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, Optional, Union, List
 
 import gymnasium as gym
 import numpy as np
@@ -27,6 +27,7 @@ from openrl.envs.vec_env import BaseVecEnv, SyncVectorEnv
 from openrl.envs.wrappers.monitor import Monitor
 from openrl.utils.callbacks.callbacks import BaseCallback, EventCallback
 from openrl.utils.evaluation import evaluate_policy
+import openrl.utils.callbacks.callbacks_factory as callbacks_factory
 
 env_wrappers = [
     Monitor,
@@ -61,7 +62,7 @@ class EvalCallback(EventCallback):
     :param eval_env: The environment used for initialization
     :param callback_on_new_best: Callback to trigger
         when there is a new best model according to the ``mean_reward``
-    :param callback_after_eval: Callback to trigger after every evaluation
+    :param callbacks_after_eval: Callback to trigger after every evaluation
     :param n_eval_episodes: The number of episodes to test the agent
     :param eval_freq: Evaluate the agent every ``eval_freq`` call of the callback.
     :param log_path: Path to a folder where the evaluations (``evaluations.npz``)
@@ -79,8 +80,12 @@ class EvalCallback(EventCallback):
     def __init__(
         self,
         eval_env: Union[str, Dict[str, Any], gym.Env, BaseVecEnv],
-        callback_on_new_best: Optional[BaseCallback] = None,
-        callback_after_eval: Optional[BaseCallback] = None,
+        callbacks_on_new_best: Optional[
+            Union[List[Dict[str, Any]], Dict[str, Any], BaseCallback]
+        ] = None,
+        callbacks_after_eval: Optional[
+            Union[List[Dict[str, Any]], Dict[str, Any], BaseCallback]
+        ] = None,
         n_eval_episodes: int = 5,
         eval_freq: int = 10000,
         log_path: Optional[str] = None,
@@ -90,13 +95,25 @@ class EvalCallback(EventCallback):
         asynchronous: bool = True,
         verbose: int = 1,
         warn: bool = True,
+        stop_logic: str = "OR",
     ):
-        super().__init__(callback_after_eval, verbose=verbose)
+        if isinstance(callbacks_after_eval, list):
+            callbacks_after_eval = callbacks_factory.CallbackFactory.get_callbacks(
+                callbacks_after_eval, stop_logic=stop_logic
+            )
 
-        self.callback_on_new_best = callback_on_new_best
-        if self.callback_on_new_best is not None:
+        super().__init__(callbacks_after_eval, verbose=verbose)
+        self.stop_logic = stop_logic
+        if isinstance(callbacks_on_new_best, list):
+            callbacks_on_new_best = callbacks_factory.CallbackFactory.get_callbacks(
+                callbacks_on_new_best, stop_logic=stop_logic
+            )
+
+        self.callbacks_on_new_best = callbacks_on_new_best
+
+        if self.callbacks_on_new_best is not None:
             # Give access to the parent
-            self.callback_on_new_best.parent = self
+            self.callbacks_on_new_best.set_parent(self)
 
         self.n_eval_episodes = n_eval_episodes
         self.eval_freq = eval_freq
@@ -140,8 +157,8 @@ class EvalCallback(EventCallback):
             os.makedirs(os.path.dirname(self.log_path), exist_ok=True)
 
         # Init callback called on new best model
-        if self.callback_on_new_best is not None:
-            self.callback_on_new_best.init_callback(self.agent)
+        if self.callbacks_on_new_best is not None:
+            self.callbacks_on_new_best.init_callback(self.agent)
 
     def _log_success_callback(
         self, locals_: Dict[str, Any], globals_: Dict[str, Any]
@@ -231,8 +248,8 @@ class EvalCallback(EventCallback):
                         f.write(f"best model reward: {mean_reward}\n")
                 self.best_mean_reward = mean_reward
                 # Trigger callback on new best model, if needed
-                if self.callback_on_new_best is not None:
-                    continue_training = self.callback_on_new_best.on_step()
+                if self.callbacks_on_new_best is not None:
+                    continue_training = self.callbacks_on_new_best.on_step()
 
             # Trigger callback after every evaluation, if needed
             if self.callback is not None:
